+++ verik-compiler/src/main/kotlin/io/verik/compiler/transform/post/StringTemplateExpressionTransformerStage.kt
@@ -23,6 +23,7 @@ import io.verik.compiler.ast.element.expression.sv.EStringExpression
 import io.verik.compiler.ast.property.ExpressionStringEntry
 import io.verik.compiler.ast.property.LiteralStringEntry
 import io.verik.compiler.common.TreeVisitor
+import io.verik.compiler.constant.ConstantBuilder
 import io.verik.compiler.main.ProjectContext
 import io.verik.compiler.main.ProjectStage
 import io.verik.compiler.target.common.Target
@@ -42,7 +43,9 @@ object StringTemplateExpressionTransformerStage : ProjectStage() {
                 Target.C_Int -> "%0d"
                 Target.C_Double -> "%f"
                 Target.C_String -> "%s"
-                Target.C_Ubit, Target.C_Sbit -> "%h"
+                Target.C_Ubit, Target.C_Sbit -> {
+                    if (type.asBitWidth(expression) < ConstantBuilder.MIN_HEX_CONSTANT_WIDTH) "%b" else "%h"
+                }
                 Target.C_Time -> "%0t"
                 else -> "%p"
             }